+++ crates/core/src/pattern/every.rs
@@ -1,49 +1,21 @@
 use super::{
     patterns::{Matcher, Name, Pattern},
     resolved_pattern::ResolvedPattern,
-    variable::VariableSourceLocations,
     State,
 };
-use crate::{context::Context, pattern_compiler::CompilationContext, resolve};
-use anyhow::{anyhow, Result};
+use crate::{context::Context, resolve};
+use anyhow::Result;
 use im::vector;
 use marzano_util::analysis_logs::AnalysisLogs;
-use std::collections::BTreeMap;
-use tree_sitter::Node;
 
 #[derive(Debug, Clone)]
 pub struct Every {
-    pub(crate) pattern: Pattern,
+    pub pattern: Pattern,
 }
 
 impl Every {
     pub fn new(pattern: Pattern) -> Self {
         Self { pattern }
-    }
-
-    pub(crate) fn from_node(
-        node: &Node,
-        context: &CompilationContext,
-        vars: &mut BTreeMap<String, usize>,
-        vars_array: &mut Vec<Vec<VariableSourceLocations>>,
-        scope_index: usize,
-        global_vars: &mut BTreeMap<String, usize>,
-        logs: &mut AnalysisLogs,
-    ) -> Result<Self> {
-        let within = node
-            .child_by_field_name("pattern")
-            .ok_or_else(|| anyhow!("missing pattern of pattern every"))?;
-        let within = Pattern::from_node(
-            &within,
-            context,
-            vars,
-            vars_array,
-            scope_index,
-            global_vars,
-            false,
-            logs,
-        )?;
-        Ok(Every::new(within))
     }
 }
