+++ crates/core/src/pattern_compiler/every_compiler.rs
@@ -1,0 +1,37 @@
+use super::{compiler::CompilationContext, node_compiler::NodeCompiler};
+use crate::pattern::{every::Every, patterns::Pattern, variable::VariableSourceLocations};
+use anyhow::{anyhow, Result};
+use marzano_util::analysis_logs::AnalysisLogs;
+use std::collections::BTreeMap;
+use tree_sitter::Node;
+
+pub(crate) struct EveryCompiler;
+
+impl NodeCompiler for EveryCompiler {
+    type TargetPattern = Every;
+
+    fn from_node(
+        node: &Node,
+        context: &CompilationContext,
+        vars: &mut BTreeMap<String, usize>,
+        vars_array: &mut Vec<Vec<VariableSourceLocations>>,
+        scope_index: usize,
+        global_vars: &mut BTreeMap<String, usize>,
+        logs: &mut AnalysisLogs,
+    ) -> Result<Self::TargetPattern> {
+        let within = node
+            .child_by_field_name("pattern")
+            .ok_or_else(|| anyhow!("missing pattern of pattern every"))?;
+        let within = Pattern::from_node(
+            &within,
+            context,
+            vars,
+            vars_array,
+            scope_index,
+            global_vars,
+            false,
+            logs,
+        )?;
+        Ok(Every::new(within))
+    }
+}
